backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
description = "Backend for m-dashboard"
dependencies = [
    "beanie>=2.0.0",
    "fastapi[standard]",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pymongo>=4.15.3",
    "redis>=7.0.1",
    "uvicorn[standard]",
]
